refactor(appserver): extract thread/skills/config clients
